requirements.txt
nulltype==2.3.1
plaid-python==18.3.0
python-dateutil==2.8.2
python-dotenv==1.0.0
six==1.16.0
urllib3==2.1.0
